Add setup files
